First push for semantic type parsing.
