+++ config.ini
@@ -2,6 +2,13 @@
 #MAX_SENTENCE_LENGTH: Maximum length of any given sentence in the raw text.
 MAX_SENTENCE_LENGTH = 205
 
+#SEMANTIC_ANNOTATION_FILE_PATH: Location to create annotations for semantic types.
+SEMANTIC_ANNOTATION_FILE_PATH = subtest_st
+#OVERRIDE_SEMANTIC_ANNOTATIONS: If semantic type annotations exist, should we remake them?
+OVERRIDE_SEMANTIC_ANNOTATIONS = 0
+#METAMAP_PATH: Path to metamap executable for generating semantic types.
+METAMAP_PATH = /home/lab/programs/metamap/2016/public_mm/bin/metamap
+
 #EMBEDDING_SIZE: Vector length of the word2vec embeddings used.
 EMBEDDING_SIZE = 200
 #EMBEDDING_FILE: Location and name of the word2vec embeddings.
